Replace src with src2
