+++ src/constants.ts
@@ -1,5 +1,3 @@
-export const ANY_SIZE = 'ANY';
+export const ANY_SIZE = 'any';
 
-export const DEFAULT_MIN_SIZE = 256 * 256;
-
-export const DEFAULT_USER_AGENT = 'Favicon Bot (https://www.npmjs.com/package/favicon-thief)';
+// export const DEFAULT_MIN_SIZE = 256 * 256;
