+++ src/findFaviconLinks.ts
