+++ src/findIcons.ts
@@ -1,48 +1,29 @@
-import probeImageSize from 'probe-image-size';
+import { RequestInit } from 'node-fetch-cjs';
+import probeImageSize, { ProbeResult } from 'probe-image-size';
 
-import { DEFAULT_USER_AGENT } from './constants';
-import { getCandidateUrls, getHtmlCandidateUrls, sortIcons } from './lib';
-import { FindIconsOptions, Icon } from './types';
+import findFaviconLinks from './findFaviconLinks';
+import getStringWithNodeFetch from './getStringWithNodeFetch';
+import { imageSizeComparator } from './lib';
 
-/**
- * Finds all icons that represent given URL.
- * Results are sorted - best first.
- * Favors vector images, square images, and large images (in that order).
- * It never throws.
- */
-const findIcons = async (url: string, options: FindIconsOptions = {}): Promise<Icon[]> => {
-  const { init } = options;
-  const htmlCandidateUrlsStack = getHtmlCandidateUrls(url).reverse();
-  const icons: Icon[] = [];
-  const headers = {
-    'User-Agent': DEFAULT_USER_AGENT,
-    ...init?.headers,
-  };
-  let htmlCandidateUrl: string | undefined;
+const findIcons = async (url: string, init?: RequestInit): Promise<ProbeResult[]> => {
+  const faviconLinks = await findFaviconLinks(url, (url) => getStringWithNodeFetch(url, init));
+  const headers = { ...init?.headers };
+  const icons: ProbeResult[] = [];
 
-  while ((htmlCandidateUrl = htmlCandidateUrlsStack.pop())) {
+  for (const faviconLink of faviconLinks) {
     try {
-      const candidateUrls = await getCandidateUrls(htmlCandidateUrl, { ...init, headers });
+      const icon = await probeImageSize(faviconLink.url, { headers });
+      icons.push(icon);
+    } catch (error) {
+      // skip this faviconLink.url
 
-      for (const candidateUrl of candidateUrls) {
-        try {
-          const icon = await probeImageSize(candidateUrl, { headers });
-          icons.push(icon);
-        } catch {
-          // skip this candidateUrl
-        }
+      if (process.env.NODE_ENV === 'test') {
+        console.error(error);
       }
-
-      // do not try another htmlCandidateUrl website if some icon has been found
-      if (icons.length > 0) {
-        return sortIcons(icons);
-      }
-    } catch {
-      // skip this htmlCandidateUrl
     }
   }
 
-  return [];
+  return icons.sort(imageSizeComparator);
 };
 
 export default findIcons;
