+++ src/getStringWithNodeFetch.ts
