+++ src/lib/getFaviconLinks.ts
