+++ src/lib/guessFaviconUrl.ts
