+++ src/lib/index.ts
@@ -1,15 +1,14 @@
-export { default as extractCandidates } from './extractCandidates';
+export { default as createGuessedLinks } from './createGuessedLinks';
 export { default as getBaseHref } from './getBaseHref';
-export { default as getCandidateUrls } from './getCandidateUrls';
-export { default as getDefaultFaviconUrl } from './getDefaultFaviconUrl';
+export { default as getFaviconLinks } from './getFaviconLinks';
 export { default as getHtmlCandidateUrls } from './getHtmlCandidateUrls';
+export { default as guessFaviconUrl } from './guessFaviconUrl';
 export { default as extractHttpEquivRefreshUrl } from './extractHttpEquivRefreshUrl';
 export { default as imageSizeComparator } from './imageSizeComparator';
 export { default as isAbsoluteUrl } from './isAbsoluteUrl';
-export { default as isGoodIcon } from './isGoodIcon';
 export { default as parseLinkSizes } from './parseLinkSizes';
 export { default as parseHttpEquivRefresh } from './parseHttpEquivRefresh';
 export { default as parseSize } from './parseSize';
-export { default as sortCandidates } from './sortCandidates';
-export { default as sortIcons } from './sortIcons';
+export { default as sortFaviconLinks } from './sortFaviconLinks';
+export { default as sortImageSizes } from './sortImageSizes';
 export { default as unique } from './unique';
