+++ src/lib/sortFaviconLinks.ts
