--- src/lib/sortIcons.ts
@@ -1,9 +1,0 @@
-import type { Icon } from '../types';
-
-import imageSizeComparator from './imageSizeComparator';
-
-const sortIcons = (icons: Icon[]): Icon[] => {
-  return [...icons].sort(imageSizeComparator);
-};
-
-export default sortIcons;
