--- src2/constants.ts
@@ -1,3 +1,0 @@
-export const ANY_SIZE = 'any';
-
-// export const DEFAULT_MIN_SIZE = 256 * 256;
