--- src2/index.ts
@@ -1,7 +1,0 @@
-export { default as findFaviconLinks } from './findFaviconLinks';
-export { default as findIcon } from './findIcon';
-export { default as findIcons } from './findIcons';
-export { default as getStringWithNodeFetch } from './getStringWithNodeFetch';
-export * from './constants';
-export * from './lib';
-export * from './types';
