--- src2/lib/index.ts
@@ -1,14 +1,0 @@
-export { default as createGuessedLinks } from './createGuessedLinks';
-export { default as getBaseHref } from './getBaseHref';
-export { default as getFaviconLinks } from './getFaviconLinks';
-export { default as getHtmlCandidateUrls } from './getHtmlCandidateUrls';
-export { default as guessFaviconUrl } from './guessFaviconUrl';
-export { default as extractHttpEquivRefreshUrl } from './extractHttpEquivRefreshUrl';
-export { default as imageSizeComparator } from './imageSizeComparator';
-export { default as isAbsoluteUrl } from './isAbsoluteUrl';
-export { default as parseLinkSizes } from './parseLinkSizes';
-export { default as parseHttpEquivRefresh } from './parseHttpEquivRefresh';
-export { default as parseSize } from './parseSize';
-export { default as sortFaviconLinks } from './sortFaviconLinks';
-export { default as sortImageSizes } from './sortImageSizes';
-export { default as unique } from './unique';
