--- src2/lib/parseSize.ts
@@ -1,15 +1,0 @@
-import type { ImageSize } from '../types';
-
-const parseSize = (size: string): ImageSize | null => {
-  const [widthString, heightString] = size.split(/x/i);
-  const width = parseInt(widthString, 10);
-  const height = parseInt(heightString, 10);
-
-  if (Number.isNaN(width) || Number.isNaN(height)) {
-    return null;
-  }
-
-  return { width, height };
-};
-
-export default parseSize;
