--- src2/lib/unique.ts
@@ -1,5 +1,0 @@
-const unique = <T>(array: T[]): T[] => {
-  return Array.from(new Set(array));
-};
-
-export default unique;
